+++ src/index.ts
@@ -3,3 +3,16 @@ export * from './KeyTransfer'
 export * from './AccessProofCondition'
 export * from './AccessProofTemplate'
 export * from './utils'
+export * from './ConditionTemplate'
+export * from './AccessProofTemplate.serviceAgreementTemplate'
+export * from './NFT721AccessProofTemplate'
+export * from './NFT721AccessProofTemplate.serviceAgreementTemplate'
+export * from './NFT721SalesWithAccessTemplate'
+export * from './NFT721SalesWithAccessTemplate.serviceAgreementTemplate'
+export * from './NFTAccessProofTemplate'
+export * from './NFTAccessProofTemplate.serviceAgreementTemplate'
+export * from './NFTSalesWithAccessTemplate'
+export * from './NFTSalesWithAccessTemplate.serviceAgreementTemplate'
+export * from './ProofTemplate'
+export * from './Service'
+export * from './utils'
