+++ test/ConsumeAssetProof.test.ts
@@ -10,8 +10,7 @@ import {
 } from '@nevermined-io/nevermined-sdk-js'
 import { assert } from 'chai'
 import { decodeJwt } from 'jose'
-import { Dtp } from '../src/Dtp'
-import { makeKeyTransfer } from '../src/KeyTransfer'
+import { Dtp, makeKeyTransfer } from '../src'
 import { config } from './config'
 import { cryptoConfig, getMetadataForDTP, sleep } from './utils'
