+++ test/KeyTransfer.test.ts
@@ -1,5 +1,5 @@
 import { assert } from 'chai'
-import { makeKeyTransfer } from '../src/KeyTransfer'
+import { makeKeyTransfer } from '../src'
 
 describe('KeyTransfer', () => {
   let keyTransfer
@@ -29,22 +29,19 @@ describe('KeyTransfer', () => {
     it('hashing works', async () => {
       assert.equal(
         await keyTransfer.hashKey(data),
-        '0x0e7f3c2e154c0793d96cad8b90862d41f58f6e526b42241bcd0b0ccfca8ba4f2'
+        '0x0e7f3c2e154c0793d96cad8b90862d41f58f6e526b42241bcd0b0ccfca8ba4f2',
       )
     })
     it('can get the secret using ecdh', async () => {
       assert.equal(
         await keyTransfer.ecdh(providerK, buyerPub),
-        await keyTransfer.ecdh(buyerK, providerPub)
+        await keyTransfer.ecdh(buyerK, providerPub),
       )
     })
     it('can encrypt and decrypt the key', async () => {
       const mimcSecret = await keyTransfer.ecdh(providerK, buyerPub)
       const cipher = await keyTransfer.encryptKey(data, mimcSecret)
-      assert.equal(
-        data.toString('hex'),
-        keyTransfer.decryptKey(cipher, mimcSecret).toString('hex')
-      )
+      assert.equal(data.toString('hex'), keyTransfer.decryptKey(cipher, mimcSecret).toString('hex'))
     })
     it('proving works', async () => {
       await keyTransfer.prove(buyerPub, providerPub, providerK, data)
