+++ test/utils.test.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from 'fs'
-import { decrypt, encrypt, aes_decryption_256, aes_encryption_256 } from '../src/utils'
+import { decrypt, encrypt, aes_decryption_256, aes_encryption_256 } from '../src'
 import { assert } from 'chai'
 
 describe('utils', () => {
@@ -12,7 +12,7 @@ describe('utils', () => {
       provider_password: 'secret',
       provider_key: readFileSync('./test/data/provider.json').toString(),
       provider_rsa_public: readFileSync('./test/data/rsa_pub_key.pem').toString(),
-      provider_rsa_private: readFileSync('./test/data/rsa_priv_key.pem').toString()
+      provider_rsa_private: readFileSync('./test/data/rsa_priv_key.pem').toString(),
     }
   })
   it('should encrypt and decrypt using RSA', async () => {
@@ -25,7 +25,7 @@ describe('utils', () => {
     const msg_ = await decrypt(config, result, 'PSK-ECDSA')
     assert.equal(msg_, msg)
   })
-  it('should encrypt and decrypt using password', async() => {
+  it('should encrypt and decrypt using password', async () => {
     const passwd = 'eMo4geeoHD6Buv3lwT1ljlVc7SHMcX9zx4uuEdEpj2M'
     const msg = readFileSync('./test.data').toString()
     const ctext = aes_encryption_256(msg, passwd)
